Pop Up Alerts
Added pop up alerts for the addPlayerButton for when user trying to put
in an invalid name (empty name), duplicates, and when the user tries to
add when the game lobby is already full.

Added pop up alerts for the deletePlayerButton for when the user
attempts to delete when nothing is selected.